Wrapping libfes using cython, so we can call it directly from python.
This adds a uptide.fes2012 module with a Fes class that implements the core functionality
of the libfes library. The libfes library has to be build with -fPIC:

  CFLAGS=-fPIC -O2 ./configure && make

To install uptide with direct uptide support:

  FES_DIR=<path-to-fes-src> python setup.py install --with-fes --prefix=$(HOME)

NOTE: that this only recomplies fes2012.c it does not rerun cython (to avoid
adding a cython dependency). To regenerate fes2012.c from fes2012.pyx, run:

  cython fes2012.pyx